Finishing te drawing lines script
the script now has boundaries to spawn the tong.
it doesnt allow the user to swipe higher than the first touch.
the script is called as an event from the event system with the head Player actions script.
the tong comes back to its mouth after reaching the same distance of the user's input.

fix the issue of moving the frog makes the tong to move as well, the tong must keep its path even if the frog has changed its position