correcting he bug where the frog faces backwards
the frog was facing back even with the limit, now the frog will face the target only when the tong is not in its mouth.
the line was showing below the limit even if it was out of boundaries, now the line will show only if it is in its boundaries.

Fix the issue where the tip of the tong has a gap between its body and itself