Finishing the functionalities and improvements
the frog functionalities are ready, the frog can move side to side, can spawn its tong, it has specific boundaries and bools to control the behavior of the tong,
the code has documentation and tooltips to guide the developer,
the code is clean, I have deleted redundancies and scripts and components that are not required right now.

the code is ready to merge to continue with the art, enemies, animations and effects.