fixing the body movement
now the body movement is smooth, it takes some time to travel from the original position to the desired position. I am using a coroutine to control the movement and the frog can only move if it is not currently in movement.

fixing names of variables of the tong behavior script, fixing some problems with the logic with the movement with the tong. now the tong is smooth.