setting events for the score system
the creatures pass their score to the score system manager.
addig a combo system, where the tong will check the number of items caught, if the count is zero the max combo will be restarted to 0, but if it is greater than zero, the maxcombo will increase by one, multiplying the score of the creatures caught.

I need to fix the heads rotation, the frog rotates the head when is attacking. and clean the spawningTongCoroutine() because I can create one function and call it twice.