finishing the basic model of the score slider
adding the textmeshpro to add the score system.
importing the text mesh pro to the project.
fixing the bug when the users touches the screen before the tong cames back to its mouth.
sharing the score from the game manager with the UI components.
fixing the combos, I can deactivate them.

I need to pass the combo counter to the tong to check if the tong caught something.
fix the line because it is behind of the frog