changing the way the frogs moves side to side
now the frogs moves like the game surfer surfer that if the players swipes left the character will move left.
the frogs head wont move if the player touches the screen if the frog is attacking and the frog is moving side to side.
correcting a bug when the tong is coming back the pivot can change if the frogs moves side to side, for that reason the finalPos in the tongBehavior script is overwritten over and over again when the tong is comming back