correcting the responsive design of the background
the background now can adjust depending on the screen size,
I utilize one package called Letterbox, which help me to add 2 cameras to the scene, one of them is the playable area I can select what is my playable area in this case is 9:16 of the game,
and the second is to control what you can see outside the area of the game.

Because I decided to put a second camera with the scrolling parallax effect, I need to adjust the second camera depending on the screen size, if I dont adjust the camera the background will be cut, or it will not match with the bg of the first camera.

I added 2 images for the background, one is the floor and the other is the grass, the BG is implemented right now but I want to polish my art, in the future, I just need to change the sprite and that is it.

I deleted code that I dont need anymore, such as the old parallax effect, the old background art