finishing the tong behavior
the tong is finished, I just need to create a nice sprite for the line and the target point.
the frog can move when attacking, and can attack when moving.

I got rid of the line to show when the frog will move side to side, I don't think it is necessary to display that behavior. Also, I cleaned the code and fixed a bug on the script TongBehaviour line 52 to 57