juice articles and counter to start game
creating a prefab that will be the particle system for the juices, instantiating the prefab when the frog catches a creature, the prefab will be instantiated and after couple second it will go to the frog.

moving the scripts to their respective folder to have a good organization.

improving the count down system, now I added the LuckiestGuy font and changed the color, now the UI game object contains the script and it will set active to false the count down game object.

Fixing the movement of the frog, I didn't add the time.deltatime to the frog, creatures, juice particles, and tong movement, before the movement was different from my computer and cellphone

hiding the pause and stats ui when the game is over.